dev-requirements.txt
# File generated by ./tools/refreeze-requirements.sh
astroid==4.0.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
coverage[toml]==7.13.4
cryptography==46.0.7
dill==0.4.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pygments==2.20.0
pylint==4.0.5
pytest==9.0.3
pytest-cov==7.1.0
pyyaml==6.0.3
requests==2.33.1
rsa==4.9.1
tomlkit==0.14.0
typing_extensions==4.15.0
urllib3==2.6.3
watchdog[watchmedo]==6.0.0
zipp==3.23.1
